% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{acc_est}
\alias{acc_est}
\title{Estimation of Accuracy Measures}
\usage{
acc_est(Z, beta, lambda0)
}
\arguments{
\item{Z}{baseline covariates}

\item{beta}{regression coefficients}

\item{lambda0}{baseline hazards}
}
\description{
Estimation of Accuracy Measures
}
