% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dph.R
\name{dph_fit}
\alias{dph_fit}
\title{Discrete Proportional Hazards (DPH) Model}
\usage{
dph_fit(t, d, X)
}
\arguments{
\item{t}{discrete time (vector)}

\item{d}{event indicator (vector)}

\item{X}{covariates (matrix)}
}
\description{
Discrete Proportional Hazards (DPH) Model
}
