% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik.R
\name{loglik}
\alias{loglik}
\title{Log-likelihood}
\usage{
loglik(beta, lambda0, theta, phi, t0, d0, X)
}
\arguments{
\item{beta}{regression coefficients}

\item{lambda0}{baseline hazards}

\item{theta}{sensitivities}

\item{phi}{specificities}

\item{t0}{observed event time}

\item{d0}{observed event indicator}

\item{X}{baseline covaraites}
}
\description{
Log-likelihood
}
